To quote the Wikipedia entry for the University of Duisburg-Essen, "In 2003 the Gerhard-Mercator-University and the University of Essen merged to the University of Duisburg-Essen.", so the two entries for Thomas Dreibholz are probably for the same person; merge them.
teluna.org is the site for a Joost Damad and an Isabelle Marien, and
following the links to his blog indicates that he's a Debian user and at
least uses openMSX.  A search for Joost Yervante Damad also finds a
recommendation to accept a Joost Yervante Damad as a Debian developer;
he says he maintains openMSX and is "a software developer and integrator
for a large multinational".  My guess is that said large multinational
is Siemens, so I'm assuming the two Joost Yervante Damad entries are for
the same person.

That leaves the two Thomas Palmers; they might be the same person, but
it's conceivable that they're not, so I'll do a bit more digging before
combining those entries.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@28632 f5534014-38df-0310-8fa8-9805f1628bb7