From Florian Forster via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=4002
Version 4.7 of collectd introduces cryptographic features for the network
protocol. The current version of the collectd dissector does not know how to
handle these segments and flags them as unknown.

The following two patches add simple support for the TYPE_SIGN_SHA256 and
TYPE_ENCR_AES256 segments. They display the plain text components of
these segments but do not try to verify the signatures or decrypt the encrypted
payload.

Changes by me:
	- replace tvb_length_remaining() with tvb_reported_length_remaining()
	- replace proto_tree_add_protocol_format() with tvb_tree_add_text()


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@29913 f5534014-38df-0310-8fa8-9805f1628bb7