The dcerpc dissectors used to walk the list of items upward parent by parent in order to push teh display of extra interesting fields in the packet to higher up in the decode tree. This was useful for making sure that things like DomainNames etc are clearly visible without having to drill down 500 layers of NDR.
This code used to just blindly walk the indicated number of parents, and 
then attach the string to that item.
This relied on the "unsafe" assumption that the topmost item would have 
pointer point to itself, so this was "safe".

This is no longer safe since the root object in the tree now has NULL as 
parent, and thus some of these dcerpc interfaces can now cause a SEGV 
trying to dereference NULL->parent.


I added a macro to safely walk to the parent object, or remain at the 
current object if parent is NULL.



This was a serious bug, where dcerpc traffic could cause a SEGV.
Please merge into all stable versions.






git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@30208 f5534014-38df-0310-8fa8-9805f1628bb7