Bug 3597 - implicit octet string that is constructed causes PRES/FTAM dissect failure
Introduced some state to remember last dissected Tag/Length so that they can be recalled if an IMPLICIT tag is encountered and stripped. This allows its to be determined if the value has a constructed value - and so can be reassembled.

In this case, it is a IMPLICIT constructed OCTET STRING at the presentation layer.

Many thanks to Fred Gruman for identifying - and apologies for the delay in commiting.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@33048 f5534014-38df-0310-8fa8-9805f1628bb7