From Reinhard Speyerer via bug 4779:
From reading the rawshark(1) manpage my assumption was that rawshark
could be used like
$ /usr/bml/bin/rawshark -s -r test.pcap -d encap:EN10MB ...

However rawshark either expects the -r argument to be -
(read from stdin) or a pipe which results in the following error
message:

rawshark: ".../test.pcap" is neither an interface nor a pipe

The proposed rawshark.pod patch updates the -r description to
the implemented rawshark functionality.

The patch also applies to the current SVN version.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@33063 f5534014-38df-0310-8fa8-9805f1628bb7