No need to have the same separator line twice, once with the new packet list and once with the old packet list.
Have Boolean variables for some of the predicates in
set_menus_for_selected_packet(), to clarify what they're testing.

Clean up indentation.

Add some comments for various issues.

(FYI, what "Restore the functionality of <Ctrl>A and <Ctrl>X to the
filter textbox." means, in the comment for the previous checkin, "don't
use <Ctrl>A or <Ctrl>X for our own purposes, so they work the same way
they do in other applications using GTK+"; the checkin changed some menu
items not to use <Ctrl>A and <Ctrl>X as accelerators.)


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@33823 f5534014-38df-0310-8fa8-9805f1628bb7