Inspired by https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3548 :
Add the option to check hf entries for various problems.  This is disabled by
default for the moment as it's quite noisy.

The regexps may need some additional fine tuning (or maybe simplification),
but they seem to work pretty well.

Arguably this might make more sense in checkhf.pl but it was easier to integrate
into this script.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@34240 f5534014-38df-0310-8fa8-9805f1628bb7