From Thomas Boehne via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5324 :
As mentioned before, the heuristics used by packet-adwin-config were weak for
UDP packets. This patch creates a stronger heuristic by checking if one of the
MAC addresses in the UDP packets is in the range assigned to the manufacturer
of the devices that use this protocol (or if it is a broadcast).

That should be relatively safe to do because the protocol is rarely used in
routed networks.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@34654 f5534014-38df-0310-8fa8-9805f1628bb7