From Neil Piercy via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3301 :
The packet-sccp.c has a bug in the declared valid ranges of the SSN and DPC
values in the user table used to match to a subdissector. The SSN range is 16
bits rather than 8 (not really an issue) but the DPC range is 16 bits rather
than 24 - so many traces cannot be matched by this table.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@35097 f5534014-38df-0310-8fa8-9805f1628bb7