Use reported length (instead of the captured length) when dissecting error causes (should have been in rev 35366).
When generating TVB subsets, limit the subset's backing and reported lengths to
the (captured) TVB length and the reported TVB length, respectively.

This allows us to dissect most of the packet in
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5546
before asserting out.

It also yields similar better behavior when the capture is limited by a
snapshot length.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@35368 f5534014-38df-0310-8fa8-9805f1628bb7