Add support for passing NTP times to proto_tree_add_item() by specifying an encoding of ENC_TIME_NTP.
This increases the number of decimal places shown for NTP times (from 6 to 9),
so round the value to the nearest microsecond.  (I can't tell if NTP times are
ever more precise than a microsecond--this rounding is mainly to be closer to
the old behavior.)

Use proto_tree_add_item() for some NTP times.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@35840 f5534014-38df-0310-8fa8-9805f1628bb7