The lack of _WITH_PHDR in WTAP_ENCAP_BLUETOOTH_H4 means there's no pseudo-header, and hence there's no direction indication. Don't set pinfo->p2p_dir for it.  Use WTAP_ENCAP_BLUETOOTH_H4_WITH_PHDR, not WTAP_ENCAP_BLUETOOTH_H4, for capture files where we have the direction.
Don't assume pinfo->p2p_dir is either P2P_DIR_SENT or P2P_DIR_RECV when
setting the info column in various Bluetooth dissectors; it might be
unknown.

In the HCI H4 dissector, put the direction into the info column
regardless of whether we have a type match or not; the dissectors for
HCI packet types appear to assume it's been set (as they put a blank at
the beginning of the stuff they append to the direction).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@35933 f5534014-38df-0310-8fa8-9805f1628bb7