When reassembling fragments, don't stop looking at fragments just because the current fragment pushes us past the reassembled size: it may be that the current fragment is a duplicate/retransmission and will be ignored.
Also, if we detect a conflict between a previous and the current fragment,
flag the current (conflicting) fragment as FD_OVERLAPCONFLICT.  Do *not* flag
the fragment that got us into the reassembly routine (probably the final
fragment): it is not (may not be) the guilty fragment.

Clean up some spacing.

Also add reassembly tests for duplicate/retransmitted fragments.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@36131 f5534014-38df-0310-8fa8-9805f1628bb7