From Jakub Zawadzki: speed up random access to gzipped files, as per the zran.c example in the zlib source.
This means that problems in the file's contents might not be reported
when a packet is read, as long as there's no problem in the contents of
the file up to the last bit of compressed data for the packet; we now
check for errors after finishing the sequential read of the file, at
least in some programs, so that shouldn't be an issue (the other
programs need to be changed to do so as well).  This is necessary in
order to be able to read all the packets we saw in the sequential pass;
it also lets us get a few more packets from truncated files in some
cases.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@36577 f5534014-38df-0310-8fa8-9805f1628bb7