From Solomon Peachy:Fix "Missing indication that we recognize the Radiotap MCS header field."
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5827

The radiotap header has a pile of optional fields, and uses a bitmask field to
indicate which fields are present.  Wireshark marks the fields it decodes, and
then decodes the payload.

The recently-added MCS field is decoded now, but the bit is not displayed as
recognized. This was presumably an oversight, as this bit is recognized and the
field decoded properly.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@36632 f5534014-38df-0310-8fa8-9805f1628bb7