Instead of having GCC and CLANG variables, just have ac_supports_gcc_flags and ac_supports_W_linker_passthrough flags, the first of which, for now, we set for GCC and clang, and the latter of which we set for GCC, clang, and xlc (probably true for some other compilers as well).
Rename AC_WIRESHARK_GCC_LDFLAGS_CHECK to AC_WIRESHARK_LDFLAGS_CHECK, as
it's not checking for anything GCC-specific.  (Leave
AC_WIRESHARK_GCC_CFLAGS_CHECK unrenamed for now, as the flags we test
with it are originally GCC flags that clang also supports for GCC
compatibility.)

Fix some string-equality tests to use = rather than ==; the former is
what the test/[ command uses.

Don't turn on "-no-cpp-precomp" for clang - it whines if you do.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@36731 f5534014-38df-0310-8fa8-9805f1628bb7