new_packet_list_find_row_from_data() is always used to select a packet, so get rid of the select_flag argument, and rename it new_packet_list_select_row_from_data().
It's also always passed a frame_data *, so make its argument a
frame_data *.

Its return value is used only to detect whether the packet was found in
the display or not, so make it a gboolean.  Check it in *all* cases
where it's called, and change the dialog message a bit (the most likely
cause is that the user cancelled a redissection of the packets, so not
all packets in the capture file are in the display.

Also, in the find case, pass it the new packet we found.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@36839 f5534014-38df-0310-8fa8-9805f1628bb7