From Loránd Jakab via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5854 :
The Locator/ID Separation Protocol [1] is being standardized within the IETF,
and it is nearing RFC status (pending security review). I have been maintaining
a dissector patch for about a year, see [2]. Feedback received indicates that,
among others, it is widely used by the developers of a large router vendor,
without issues.

In January I submitted the dissector for data plane packets as bug #5602, which
was committed as r35615. The patch attached to this bug adds support for
dissection of control plane packets.

[1] http://tools.ietf.org/html/draft-ietf-lisp
[2] http://lisp.ccaba.upc.edu/wireshark/


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@36845 f5534014-38df-0310-8fa8-9805f1628bb7