Create a new frame_data_sequence data type; it represents a dense sequence of frame_data structures, indexed by the frame number.  Extract the relevant bits of the capture_file data structure and move them to the frame_data_sequence, and move the relevant code from cfile.c and tweak it to handle frame_data_sequence structures.
Have a possibly-null pointer to a frame_data_sequence structure in the
capture_file structure; if it's null, we aren't keeping a sequence of
frame_data structures (we don't keep that sequence when we're doing
one-pass processing in TShark).

Nothing in libwireshark should care about a capture_file structure; get
rid of some unnecessary includes of cfile.h.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@36881 f5534014-38df-0310-8fa8-9805f1628bb7