SVN #37419 fixed an OSX compile issue and reverted SVN #37408.
The actual cause of an OSX compile problem: The text   [^[:blank:]]   in a comment.

SVN #37408 was reverted because a period (or most other characters)
in a RE character class don't need to be escaped.

This patch adds some replacement text in the spot in the comment
originally occupied by the text    [^[:blank:]]



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@37421 f5534014-38df-0310-8fa8-9805f1628bb7