Back out parts of 38112, as pointed out by Stig Bj\370rlykke:
The changes in profile_dlg.c, uat_gui.c and new_packet_list.c does not
work.  The tooltip is not displayed, and it's not possible to get the
context menu from a column header in the packet list.
...
I guess gtk_tree_view_column_get_widget() does not return
column->button, which is the widget we use.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@38117 f5534014-38df-0310-8fa8-9805f1628bb7