Set eth.fcs_len in the pseudo-header for Ethernet packets.
Set the pseudo-header when doing the sequential read as well as when
doing random reads.

When writing packets to a CommView file, use a slightly less contorted
way to get the year/month/day/hour/minute/second values.

commview_dump() uses the pseudo_header argument; don't mark it as
unused.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@38833 f5534014-38df-0310-8fa8-9805f1628bb7