Use ENC_ values in proto_tree_add_item() calls.
Get rid of the code that attempts to deal with the FCS; if the
link-layer dissector hasn't removed the FCS, either:

	1) the link-layer dissector is buggy and should be fixed;

or

	2) the link-layer dissector doesn't know whether there's an FCS
	   and you have to tell it by setting a preference (and if there
	   is no such preference, one needs to be added).

The code in question was causing errors when the link-layer dissector
was removing the FCS when it's known to be present, as it should do.

Get rid of 4-space tabs in favor of spaces.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@39158 f5534014-38df-0310-8fa8-9805f1628bb7