Clean up white space.
The appropriate encoding for a string is ENC_BIG_ENDIAN or
ENC_LITTLE_ENDIAN, combined with a character encoding value; it is not
ENC_NA, as the character encoding is definitely applicable, and the byte
order is also applicable if it's a counted string or a UTF-16 or UCS-2
string.  Assume ASCII for now.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@39187 f5534014-38df-0310-8fa8-9805f1628bb7