From Jouni Malinen via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6422 Clean up IEEE 802.11 dissector - fixed fields
The fixed field identifiers were defined to have specific values. However, this is used only within the parser and does not correspond to any specific packet field. As such, there is no need for the specific values to be maintained and an enum makes it simpler to add and remove these fields as needed.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@39315 f5534014-38df-0310-8fa8-9805f1628bb7