Added a private hash table to transport string values between dissectors. This works between C and Lua.
In C the pinfo.private_table pointer must be initialized using
g_hash_table_new (g_str_hash, g_str_equal);

In Lua the values are available using pinfo.private.<key>, and the
table is created automatically on first usage.  It's possible to use
this datatypes: nil, boolean, number and string, but every value
is converted to string so numbers must be converted using tonumber()
on usage.  Boolean is either nil or an empty string.

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@39461 f5534014-38df-0310-8fa8-9805f1628bb7