If we set "column.hidden" on the command line but don't set "column.format" in a later argument, make sure we mark the appropriate existing columns as hidden.
As part of cleaning up the infrastructure for doing this, store the COL_
code for a column in a fmt_data structure rather than a pointer to the
raw column format string; this simplifies some code.

Have a routine to return whether a column is visible, not hidden, as
it's used to set a "visible" flag.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@39822 f5534014-38df-0310-8fa8-9805f1628bb7