Support nanosecond-resolution time for NetMon 2.x format (it's only 100-nanosecond resolution, but that's still better than microsecond resolution).
For NetMon 1.x format, only claim to support millisecond resolution, as
that's all you get.

Fix handling of negative time deltas in NetMon 2.x format.

When writing a NetMon file, trim the time of the first packet to
millisecond precision to get the capture start time, so that the start
time written to the file (which has millisecond precision) is the same
as the start time used to calculate the deltas written to the packet
headers.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@39886 f5534014-38df-0310-8fa8-9805f1628bb7